add addon check in user-server
